Adversarial data-lens batch 1: table types and spellings
All verified MATCH + CHECK-GREEN:
- SearchTarget.c: SearchSight field 0 renamed near_distance ->
  sight_distance (it is the overall seeing limit; beyond it SR_UNSEEN)
  and the row gloss un-swapped ({sight 16000, clear 10000, gone
  20000}); per-field role comments added.
- StageEndScreen.c: the next-stage pick is a plain
  StageOrder[uid + 1] subscript -- retail's lbu is cc1 narrowing the
  halfword load through the u8 StageNo store, so the (u8 *) byte-walk
  and the StageEndNextStageOffset inline helper were decompiler
  artifacts; both removed.
- PlayMusicFormID.c: the music-table probes drop the redundant (s32)
  index casts (i + (s32)p keeps the measured index-first addu; plain
  p[i] flips the operands -- noted in the header).
- CheckCheatCodes.c: CheatSeq matches its sibling ForbiddenCommand as
  s16 (both memcmp'd against s16 pad records).
- SetupTelop.c: TelopFont is s16, agreeing with the s16 *font consumer
  and Krom2RawAdd's return type.
- proc_misc_bonfire_.c / spawn_damage_effect_.c: the two SVECTOR
  constants shed their u8[] element type (unknown ARRAY SIZE is the
  measured requirement, the element type was not) -- plain
  svec[0] copies now, matching every sibling constant.

+++ src/main.exe/PlayMusicFormID.c
@@ -35,7 +35,8 @@
  * its low half only after the sentinel branch. Assigning `j` after the match
  * guard lets the scheduler move it into that guard's delay slot and reuse
  * $v0, while the signed integer pointer sums preserve the target's
- * index-first `addu` operand order. The nested one-shot fence supplies the
+ * index-first `addu` operand order (measured: plain p[i] flips the
+ * addu operands; the minimal `*(u8 *)(i + (s32)p)` spelling is exact). The nested one-shot fence supplies the
  * loop weight needed for the retail register colouring without emitting code.
  */
 
@@ -78,7 +79,7 @@ void PlayMusicFormID(s32 id)
         {
             do
             {
-                if (*(u8 *)((s32)i + (s32)p) == MusicNo)
+                if (*(u8 *)(i + (s32)p) == MusicNo)
                 {
                     goto found;
                 }
@@ -86,12 +87,12 @@ void PlayMusicFormID(s32 id)
         } while (0);
         j = i + 1;
         i = j;
-        if (*(u8 *)((s32)j + (s32)p) != flag)
+        if (*(u8 *)(j + (s32)p) != flag)
         {
             goto search;
         }
     found:
-        if (*(u8 *)((s32)i + (s32)p) != 0xFF)
+        if (*(u8 *)(i + (s32)p) != 0xFF)
         {
             MusicNo = i;
         }
